Move `IncomingBody` into the `http::body` module.
Move `IncomingBody` and related code out of the `response` module and into
the `body` module, and prepare it to be used for incoming requests as well
as incoming responses.

This is in preparation for #34.
